searchapp/requirements/requirements.txt
appnope==0.1.0
asgiref==3.2.10
backcall==0.2.0
certifi==2020.6.20
decorator==4.4.2
Django==3.1.1
django-cors-headers==3.5.0
django-elasticsearch-dsl==7.1.4
django-environ==0.4.5
djangorestframework==3.11.1
elasticsearch==7.9.1
elasticsearch-dsl==7.3.0
importlib-metadata==1.7.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
Markdown==3.2.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.6.0
Pygments==2.7.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
zipp==3.1.0
